refactor(ppp): split the 3.9k-line ppp_clas.cpp into per-stage TUs
- ppp_clas_internal.hpp: former anonymous-namespace helpers (variance
  models, dump writers, MRTKLIB parity observation lookup) plus the
  slip-state helpers and cross-TU declarations
- ppp_clas_slip.cpp: detectClasCycleSlips / syncSlipState /
  markSlipCompensationFromAmbiguities
- ppp_clas_measurements.cpp: buildEpochMeasurements
- ppp_clas_update_diag.cpp: mrtklibPostfitRowsValid,
  updateObservedAmbiguities, logUpdateSummary
- ppp_clas.cpp keeps the epoch pipeline (OSR selection, preparation,
  prediction, ambiguity shifts, Kalman update, finalize)

Verified: ctest 71/71, byte-identical rtk_positioning output on PPC
tokyo/run1, MADOCALIB parity suite 17/17.
